requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
bcrypt==4.2.1
python-jose==3.3.0
requests==2.32.3
pydantic==2.10.4
sqlalchemy==2.0.36
psycopg2-binary
python-dotenv==1.0.1
loguru
pydantic[email]
passlib
python-multipart
PyPDF2
python-docx
websocket-client
boto3
pyjwt
numpy
firebase-admin